utils: Move rewrite_uri to utils
